requirements.txt
langchain==0.1.14
python-dotenv==1.0.1
langchain-experimental==0.0.56
langchain-openai==0.1.3
pandas==2.2.2
tabulate==0.9.0
streamlit==1.33.0
